rename util > common and move another dir
