+++ src/common/filteringStream.ts
