+++ src/common/isEmpty.ts
