+++ src/common/isProd.ts
@@ -1,3 +1,3 @@
-import { EnvironmentEnum } from "../../config/env";
+import { EnvironmentEnum } from "../config/env";
 
 export default () => process.env.NODE_ENV === EnvironmentEnum.production;
