+++ src/frontend/get-github-configuration.ts
@@ -1,7 +1,7 @@
 import JWT from "atlassian-jwt";
 import { Installation, Subscription } from "../backend/models";
 import { NextFunction, Request, Response } from "express";
-import { getJiraMarketplaceUrl } from "../util/getUrl";
+import { getJiraMarketplaceUrl } from "../common/getUrl";
 import enhanceOctokit from "../config/enhance-octokit";
 import app from "../worker/app";
 import { getInstallation } from "./get-jira-configuration";
