+++ src/github/push.ts
@@ -1,6 +1,6 @@
 import { enqueuePush } from "../transforms/push";
 import issueKeyParser from "jira-issue-key-parser";
-import { isEmpty } from "../jira/util/isEmpty";
+import { isEmpty } from "../common/isEmpty";
 import { Context } from "probot/lib/context";
 
 export default async (context: Context, jiraClient) => {
