+++ src/jira/client/index.ts
@@ -1,8 +1,8 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { Installation, Subscription } from "../../backend/models";
 import getAxiosInstance from "./axios";
-import { getJiraId } from "../util/id";
-import isProd from "../util/isProd";
+import { getJiraId } from "../../common/id";
+import isProd from "../../common/isProd";
 import { AxiosInstance, AxiosResponse } from "axios";
 import Logger from "bunyan";
 import issueKeyParser from "jira-issue-key-parser";
