+++ src/transforms/branch.ts
@@ -1,7 +1,7 @@
-import { getJiraId } from "../jira/util/id";
+import { getJiraId } from "../common/id";
 import { Context } from "probot/lib/context";
 import issueKeyParser from "jira-issue-key-parser";
-import { isEmpty } from "../jira/util/isEmpty";
+import { isEmpty } from "../common/isEmpty";
 
 async function getLastCommit(context, issueKeys) {
 	const {
