+++ src/transforms/commit.ts
@@ -1,5 +1,5 @@
 import issueKeyParser from "jira-issue-key-parser";
-import { isEmpty } from "../jira/util/isEmpty";
+import { isEmpty } from "../common/isEmpty";
 import { JiraCommit, JiraCommitData } from "../interfaces/jira";
 
 function mapCommit(githubCommit, author): JiraCommit {
